spawn tetrominos on board

+++ app/src/main/java/tetris/utils/Point.java
@@ -1,27 +1,27 @@
 package tetris.utils;
 
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setX(int val) {
+    public void setX(double val) {
         x = val;
     }
 
-    public void setY(int val) {
+    public void setY(double val) {
         y = val;
     }
 }
